Conflicts:
	bin/unit/Unit.class
	bin/unit/UnitBattleManager$CombatInfo.class
	bin/unit/UnitBattleManager.class
	gitignore
	src/unit/UnitBattleManager.java

+++ src/unit/Unit.java
@@ -20,7 +20,7 @@ public class Unit {
 	public final Player owner;
 	
 	//	Moves
-	private int movementPoints;
+	private byte movementPoints;
 	
 	//	Health
 	private byte currentHealth;
